requirements.txt
streamlit
pandas
python-dotenv
pybit
ta
plotly
jinja2
requests

